Restructure test-utils package
